Restructure Appointment folder
