+++ src/components/appointment/Appointment.css
@@ -1,0 +1,106 @@
+.appointment-page {
+    margin: 0;
+    padding: 0;
+    background-image: url('../../background.jpg');
+    background-size: cover;
+    background-position: center;
+    height: 100vh;
+    display: flex;
+    align-items: center;
+    position: relative;
+    z-index: 1;
+  }
+  
+  .appointment {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .foreground-color {
+    background: rgba(25, 26, 25, 0.8);
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    z-index: -1;
+  }
+  
+  .appointment-content {
+    position: relative;
+    z-index: 1;
+    text-align: center;
+    padding: 20px;
+    color: white;
+  }
+  
+  .appointment-content h2 {
+    width: 60%;
+    margin: 0 auto;
+    border-bottom: 1px solid #9b9797;
+    padding-bottom: 15px;
+    margin-bottom: 20px;
+  }
+  
+  .appointment-form {
+    margin-top: 30px;
+    font-family: 'Roboto', sans-serif;
+    font-size: 16px;
+  }
+  
+  .appointment-form select {
+    padding: 10px 25px;
+    background-color: rgb(124, 196, 10);
+    border: 2px solid #fff;
+    border-radius: 20px;
+    margin-right: 25px;
+  }
+  
+  .appointment-form #date {
+    padding: 10px 25px;
+    background-color: rgb(124, 196, 10);
+    border: 2px solid #fff;
+    border-radius: 20px;
+    margin-right: 25px;
+  }
+  
+  .appointment-form #submit {
+    padding: 12px 27px;
+    background-color: #fff;
+    border: none;
+    border-radius: 20px;
+    cursor: pointer;
+  }
+  
+  @media (max-width: 768px) {
+    .appointment-content {
+      padding: 10px;
+    }
+  
+    .appointment-content h2 {
+      width: 90%;
+      font-size: 24px;
+    }
+  
+    .appointment-form {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
+  
+    .appointment-form select,
+    .appointment-form #date,
+    .appointment-form #submit {
+      width: 100%;
+      margin: 10px 0;
+    }
+  }
+  
+  @media (max-width: 480px) {
+    .appointment-content h2 {
+      font-size: 20px;
+    }
+  }
+
